+++ make/plugin.go
@@ -8,21 +8,39 @@ package make
 
 import (
 	"context"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
 )
 
 const pluginName = "steampipe-plugin-make"
 
-// Plugin creates this (make) plugin
 func Plugin(ctx context.Context) *plugin.Plugin {
 	p := &plugin.Plugin{
 		Name:             pluginName,
 		DefaultTransform: transform.FromCamel().NullIfZero(),
+		// SchemaMode: plugin.SchemaModeDynamic,
+		// DefaultTransform: transform.FromGo(),
+		// DefaultGetConfig: &plugin.GetConfig{
+		//	IgnoreConfig: &plugin.IgnoreConfig{
+		//		ShouldIgnoreErrorFunc: isNotFoundError([]string{
+		//			"NoSuchEntity",
+		//			"NotFoundException",
+		//			"ResourceNotFoundException",
+		//			"InvalidParameter",
+		//			"InvalidParameterValue",
+		//			"InvalidParameterValueException",
+		//			"ValidationError",
+		//			"ValidationException",
+		//		}),
+		//	},
+		// },
+		// Default ignore config for the plugin
+		// DefaultIgnoreConfig: &plugin.IgnoreConfig{
+		//	ShouldIgnoreErrorFunc: shouldIgnoreErrorPluginDefault(),
+		// },
 		ConnectionConfigSchema: &plugin.ConnectionConfigSchema{
-			NewInstance: client.ConfigInstance,
-			Schema:      client.ConfigSchema,
+			NewInstance: ConfigInstance,
+			Schema:      ConfigSchema,
 		},
 		TableMap: map[string]*plugin.Table{
 			"make_api_token":              tableApiToken(ctx),
@@ -30,6 +48,9 @@ func Plugin(ctx context.Context) *plugin.Plugin {
 			"make_data_store":             tableDataStore(ctx),
 			"make_organization":           tableOrganization(ctx),
 			"make_organization_variable":  tableOrganizationVariable(ctx),
+			"make_scenario":               tableScenario(ctx),
+			"make_scenario_dlq":           tableScenarioDlq(ctx),
+			"make_scenario_log":           tableScenarioLog(ctx),
 			"make_team":                   tableTeam(ctx),
 			"make_team_variable":          tableTeamVariable(ctx),
 			"make_user":                   tableUser(ctx),