+++ make/table_make_api_token.go
@@ -2,7 +2,6 @@ package make
 
 import (
 	"context"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
@@ -18,9 +17,10 @@ func tableApiToken(_ context.Context) *plugin.Table {
 		Columns: []*plugin.Column{
 			// Other Columns
 			{Name: "token", Type: proto.ColumnType_STRING, Description: "The API Token (masked)."},
-			{Name: "label", Type: proto.ColumnType_STRING, Description: "The friendly name of the API Token."},
+			{Name: "label", Type: proto.ColumnType_STRING, Description: "The name of the API Token."},
 			{Name: "scope", Type: proto.ColumnType_JSON, Description: "Scopes enabled for the API Token."},
 			{Name: "created", Type: proto.ColumnType_TIMESTAMP, Description: "Creation date of the API Token."},
+			{Name: "is_active", Type: proto.ColumnType_BOOL, Description: "Is the API Token currently used in make.spc?"},
 
 			// Standard Columns
 			{Name: "title", Type: proto.ColumnType_STRING, Description: StandardColumnDescription("title"), Transform: transform.FromField("Label")},
@@ -31,29 +31,28 @@ func tableApiToken(_ context.Context) *plugin.Table {
 func listApiTokens(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("listApiTokens", ctx, d, h)
 
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// prepare params
-	var config = client.NewRequestConfig("users/me/api-tokens")
-	if d.QueryContext.Limit != nil {
-		config.Pagination.Limit = int(*d.QueryContext.Limit)
-	}
+	var op = c.NewApiTokenListPaginator(int(d.RowsRemaining(ctx)))
+	for op.HasMorePages() {
+		tokens, err := op.NextPage()
+		if err != nil {
+			plugin.Logger(ctx).Error("make_api_token.listApiTokens", "request_error", err)
+			return nil, err
+		}
 
-	// fetch data
-	var result = &client.ApiTokenListResponse{}
-	err = c.Get(&config, result)
-	if err != nil {
-		plugin.Logger(ctx).Error("make_api_token.listApiTokens", "request_error", err)
-		return nil, c.HandleKnownErrors(err, "user:read")
-	}
+		// stream results
+		for _, i := range tokens {
+			i.IsActive = c.IsTokenActive(i.Token)
+			d.StreamListItem(ctx, i)
 
-	// stream results
-	for _, i := range result.ApiTokens {
-		d.StreamListItem(ctx, i)
+			if d.RowsRemaining(ctx) <= 0 {
+				return nil, nil
+			}
+		}
 	}
 
 	return nil, nil