+++ make/table_make_connection.go
@@ -2,12 +2,10 @@ package make
 
 import (
 	"context"
-	"fmt"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
+	"github.com/marekjalovec/make-sdk"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
-	"strconv"
 )
 
 func tableConnection(_ context.Context) *plugin.Table {
@@ -15,8 +13,7 @@ func tableConnection(_ context.Context) *plugin.Table {
 		Name:        "make_connection",
 		Description: "For most apps included in Make, it is necessary to create a connection, through which Make will communicate with the given third-party service according to the settings of a specific scenario.",
 		List: &plugin.ListConfig{
-			Hydrate:       listConnections,
-			ParentHydrate: listOrganizations,
+			Hydrate: listConnections,
 		},
 		Get: &plugin.GetConfig{
 			KeyColumns: plugin.SingleColumn("id"),
@@ -25,21 +22,21 @@ func tableConnection(_ context.Context) *plugin.Table {
 		Columns: []*plugin.Column{
 			// Key Columns
 			{Name: "id", Type: proto.ColumnType_INT, Description: "The Connection ID."},
-			{Name: "team_id", Type: proto.ColumnType_INT, Description: "ID of the Team that owns this Connection."},
+			{Name: "team_id", Type: proto.ColumnType_INT, Description: "ID of the Team that owns the Connection."},
 
 			// Other Columns
-			{Name: "name", Type: proto.ColumnType_STRING, Description: "The user friendly name of the Connection."},
-			{Name: "account_name", Type: proto.ColumnType_STRING, Description: "The system name of the application tied to this Connection."},
-			{Name: "account_label", Type: proto.ColumnType_STRING, Description: "The user friendly name of the application tied to this Connection."},
+			{Name: "name", Type: proto.ColumnType_STRING, Description: "The name of the Connection."},
+			{Name: "account_name", Type: proto.ColumnType_STRING, Description: "The system name of the application tied to the Connection."},
+			{Name: "account_label", Type: proto.ColumnType_STRING, Description: "The name of the application tied to the Connection."},
 			{Name: "account_type", Type: proto.ColumnType_STRING, Description: "Authentication type."},
-			{Name: "package_name", Type: proto.ColumnType_STRING, Description: "Package name, present when connection belongs to a Custom App."},
+			{Name: "package_name", Type: proto.ColumnType_STRING, Description: "Name of the Custom App to which the Connection belongs to."},
 			{Name: "expire", Type: proto.ColumnType_TIMESTAMP, Description: "When does the Connection expire?"},
 			{Name: "metadata", Type: proto.ColumnType_JSON, Description: "Metadata attached to the Connection."},
 			{Name: "upgradeable", Type: proto.ColumnType_BOOL, Description: "Can the Connection be upgraded?"},
 			{Name: "scoped", Type: proto.ColumnType_BOOL, Description: "Is the Connection scoped?"},
 			{Name: "scopes", Type: proto.ColumnType_JSON, Description: "Security scopes of the Connection.", Hydrate: getConnection, Transform: transform.FromField("Scopes").NullIfEmptySlice()},
 			{Name: "editable", Type: proto.ColumnType_BOOL, Description: "Can the Connection be edited?"},
-			{Name: "uid", Type: proto.ColumnType_STRING, Description: "UID of this Connection."},
+			{Name: "uid", Type: proto.ColumnType_STRING, Description: "UID of the Connection."},
 
 			// Standard Columns
 			{Name: "title", Type: proto.ColumnType_STRING, Description: StandardColumnDescription("title"), Transform: transform.FromField("Name")},
@@ -50,84 +47,60 @@ func tableConnection(_ context.Context) *plugin.Table {
 func getConnection(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("getConnection", ctx, d, h)
 
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// prepare params
 	var id int
 	if h.Item != nil {
-		// "scopes" column detail request
-		id = h.Item.(client.Connection).Id
+		id = h.Item.(makesdk.Connection).Id
 	} else {
-		// direct query
 		id = int(d.EqualsQuals["id"].GetInt64Value())
 	}
-	var config = client.NewRequestConfig(fmt.Sprintf(`connections/%d`, id))
-
-	// fetch data
-	var result = &client.ConnectionResponse{}
-	err = c.Get(&config, &result)
+	team, err := c.GetConnection(id)
 	if err != nil {
 		plugin.Logger(ctx).Error("make_connection.getConnection", "request_error", err)
-		return nil, c.HandleKnownErrors(err, "connections:read")
+		return nil, err
 	}
 
-	return result.Connection, nil
+	return team, nil
 }
 
 func listConnections(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("listConnections", ctx, d, h)
 
-	if h.Item == nil {
-		return nil, fmt.Errorf("parent organization not defined")
-	}
-
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// iterate over organization teams
-	var teams = h.Item.(client.Organization).Teams
-	for _, team := range teams {
-		// prepare params
-		var config = client.NewRequestConfig("connections")
-		ColumnsToParams(&config.Params, []string{"id", "name", "accountName", "accountLabel", "packageName", "expire", "metadata", "teamId", "upgradeable", "scoped", "accountType", "editable", "uid"})
-		config.Params.Set("teamId", strconv.Itoa(team.Id))
-		if d.QueryContext.Limit != nil {
-			config.Pagination.Limit = int(*d.QueryContext.Limit)
+	var op = c.NewOrganizationListPaginator(-1)
+	for op.HasMorePages() {
+		organizations, err := op.NextPage()
+		if err != nil {
+			plugin.Logger(ctx).Error("make_connection.listConnections", "request_error", err)
+			return nil, err
 		}
 
-		// fetch data
-		var pagesLeft = true
-		for pagesLeft {
-			var result = &client.ConnectionListResponse{}
-			err = c.Get(&config, result)
-			if err != nil {
-				plugin.Logger(ctx).Error("make_connection.listConnections", "request_error", err)
-				return nil, c.HandleKnownErrors(err, "connections:read")
-			}
+		for _, organization := range organizations {
+			for _, team := range organization.Teams {
+				var cp = c.NewConnectionListPaginator(int(d.RowsRemaining(ctx)), team.Id)
+				for cp.HasMorePages() {
+					connections, err := cp.NextPage()
+					if err != nil {
+						plugin.Logger(ctx).Error("make_connection.listConnections", "request_error", err)
+						return nil, err
+					}
 
-			// stream results
-			for _, i := range result.Connections {
-				d.StreamListItem(ctx, i)
-			}
+					for _, i := range connections {
+						d.StreamListItem(ctx, i)
 
-			// break both cycles if we have enough rows
-			if d.RowsRemaining(ctx) <= 0 {
-				return nil, nil
-			}
-
-			// pagination
-			var resultCount = len(result.Connections)
-			if resultCount < config.Pagination.Limit {
-				pagesLeft = false
-			} else {
-				config.Pagination.Offset += config.Pagination.Limit
+						if d.RowsRemaining(ctx) <= 0 {
+							return nil, nil
+						}
+					}
+				}
 			}
 		}
 	}