+++ make/table_make_organization.go
@@ -2,8 +2,6 @@ package make
 
 import (
 	"context"
-	"fmt"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
@@ -31,7 +29,7 @@ func tableOrganization(_ context.Context) *plugin.Table {
 			{Name: "license", Type: proto.ColumnType_JSON, Description: "Licence information and limits."},
 			{Name: "zone", Type: proto.ColumnType_STRING, Description: "Zone where the Organization exists."},
 			{Name: "service_name", Type: proto.ColumnType_STRING, Description: "Service name."},
-			{Name: "is_paused", Type: proto.ColumnType_BOOL, Description: "Is the organization paused?"},
+			{Name: "is_paused", Type: proto.ColumnType_BOOL, Description: "Is the Organization paused?"},
 			{Name: "external_id", Type: proto.ColumnType_STRING, Description: "Make private instances use the externalId parameter for security reasons."},
 
 			// Standard Columns
@@ -43,65 +41,43 @@ func tableOrganization(_ context.Context) *plugin.Table {
 func getOrganization(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("getOrganization", ctx, d, h)
 
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// prepare params
-	var id = int(d.EqualsQuals["id"].GetInt64Value())
-	var config = client.NewRequestConfig(fmt.Sprintf(`organizations/%d`, id))
-	ColumnsToParams(&config.Params, []string{"id", "name", "countryId", "timezoneId", "license", "zone", "serviceName", "isPaused", "externalId", "teams"})
-
-	// fetch data
-	var result = &client.OrganizationResponse{}
-	err = c.Get(&config, &result)
+	team, err := c.GetTeam(int(d.EqualsQuals["id"].GetInt64Value()))
 	if err != nil {
 		plugin.Logger(ctx).Error("make_organization.getOrganization", "request_error", err)
-		return nil, c.HandleKnownErrors(err, "organizations:read")
+		return nil, err
 	}
 
-	return result.Organization, nil
+	return team, nil
 }
 
 func listOrganizations(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("listOrganizations", ctx, d, h)
 
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// prepare params
-	var config = client.NewRequestConfig("organizations")
-	ColumnsToParams(&config.Params, []string{"id", "name", "countryId", "timezoneId", "license", "zone", "serviceName", "isPaused", "externalId", "teams"})
-	if d.QueryContext.Limit != nil {
-		config.Pagination.Limit = int(*d.QueryContext.Limit)
-	}
-
-	// fetch data
-	var pagesLeft = true
-	for pagesLeft {
-		var result = &client.OrganizationListResponse{}
-		err = c.Get(&config, result)
+	var op = c.NewOrganizationListPaginator(int(d.RowsRemaining(ctx)))
+	for op.HasMorePages() {
+		organizations, err := op.NextPage()
 		if err != nil {
 			plugin.Logger(ctx).Error("make_organization.listOrganizations", "request_error", err)
-			return nil, c.HandleKnownErrors(err, "organizations:read")
+			return nil, err
 		}
 
 		// stream results
-		for _, i := range result.Organizations {
+		for _, i := range organizations {
 			d.StreamListItem(ctx, i)
-		}
 
-		// pagination
-		var resultCount = len(result.Organizations)
-		if d.RowsRemaining(ctx) <= 0 || resultCount < config.Pagination.Limit {
-			pagesLeft = false
-		} else {
-			config.Pagination.Offset += config.Pagination.Limit
+			if d.RowsRemaining(ctx) <= 0 {
+				return nil, nil
+			}
 		}
 	}