+++ make/table_make_user.go
@@ -2,12 +2,9 @@ package make
 
 import (
 	"context"
-	"fmt"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
-	"strconv"
 )
 
 func tableUser(_ context.Context) *plugin.Table {
@@ -15,15 +12,14 @@ func tableUser(_ context.Context) *plugin.Table {
 		Name:        "make_user",
 		Description: "Existing Users within your account, and their attributes.",
 		List: &plugin.ListConfig{
-			Hydrate:       listUsers,
-			ParentHydrate: listOrganizations,
+			Hydrate: listUsers,
 		},
 		Columns: []*plugin.Column{
 			// Key Columns
 			{Name: "id", Type: proto.ColumnType_INT, Description: "The user ID."},
 
 			// Other Columns
-			{Name: "name", Type: proto.ColumnType_STRING, Description: "Full name of the User."},
+			{Name: "name", Type: proto.ColumnType_STRING, Description: "The name of the User."},
 			{Name: "email", Type: proto.ColumnType_STRING, Description: "User's email."},
 			{Name: "language", Type: proto.ColumnType_STRING, Description: "Environment language."},
 			{Name: "locale_id", Type: proto.ColumnType_INT, Description: "Environment locale - date and hour formats, decimal separators etc."},
@@ -46,50 +42,39 @@ func tableUser(_ context.Context) *plugin.Table {
 func listUsers(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("listUsers", ctx, d, h)
 
-	if h.Item == nil {
-		return nil, fmt.Errorf("parent organization not defined")
-	}
-
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
-	// iterate over organization teams
-	var organization = h.Item.(client.Organization)
-	var teams = organization.Teams
-	for _, team := range teams {
-		// prepare params
-		var config = client.NewRequestConfig("users")
-		config.Params.Set("teamId", strconv.Itoa(team.Id))
-		if d.QueryContext.Limit != nil {
-			config.Pagination.Limit = int(*d.QueryContext.Limit)
+	var op = c.NewOrganizationListPaginator(-1)
+	for op.HasMorePages() {
+		organizations, err := op.NextPage()
+		if err != nil {
+			plugin.Logger(ctx).Error("make_user.listUsers", "request_error", err)
+			return nil, err
 		}
 
-		// fetch data
-		var pagesLeft = true
-		for pagesLeft {
-			var result = &client.UserListResponse{}
-			err = c.Get(&config, result)
-			if err != nil {
-				plugin.Logger(ctx).Error("make_user.listUsers", "request_error", err)
-				return nil, c.HandleKnownErrors(err, "user:read")
-			}
+		for _, organization := range organizations {
+			for _, team := range organization.Teams {
+				var up = c.NewUserListPaginator(-1, team.Id)
+				for up.HasMorePages() {
+					users, err := up.NextPage()
+					if err != nil {
+						plugin.Logger(ctx).Error("make_user.listUsers", "request_error", err)
+						return nil, err
+					}
 
-			// stream results
-			for _, i := range result.Users {
-				i.OrganizationId = organization.Id
-				i.TeamId = team.Id
-				d.StreamListItem(ctx, i)
-			}
+					for _, i := range users {
+						i.OrganizationId = organization.Id
+						i.TeamId = team.Id
+						d.StreamListItem(ctx, i)
 
-			// pagination
-			var resultCount = len(result.Users)
-			if d.RowsRemaining(ctx) <= 0 || resultCount < config.Pagination.Limit {
-				pagesLeft = false
-			} else {
-				config.Pagination.Offset += config.Pagination.Limit
+						if d.RowsRemaining(ctx) <= 0 {
+							return nil, nil
+						}
+					}
+				}
 			}
 		}
 	}