+++ make/table_make_user_organization_role.go
@@ -2,8 +2,6 @@ package make
 
 import (
 	"context"
-	"fmt"
-	"github.com/marekjalovec/steampipe-plugin-make/client"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
@@ -24,7 +22,8 @@ func tableUserOrganizationRole(_ context.Context) *plugin.Table {
 			// Other Columns
 			{Name: "users_role_id", Type: proto.ColumnType_INT, Description: "The ID of the Role."},
 			{Name: "organization_id", Type: proto.ColumnType_INT, Description: "The ID of the Organization."},
-			{Name: "invitation", Type: proto.ColumnType_STRING, Description: "Is the invitation is still pending?"},
+			{Name: "invitation", Type: proto.ColumnType_STRING, Description: "Invitation string."},
+			{Name: "sso_pending", Type: proto.ColumnType_BOOL, Description: "Is SSO pending?"},
 
 			// Standard Columns
 			{Name: "title", Type: proto.ColumnType_STRING, Description: StandardColumnDescription("title"), Transform: transform.FromField("Name")},
@@ -35,40 +34,27 @@ func tableUserOrganizationRole(_ context.Context) *plugin.Table {
 func listUserOrganizationRoles(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	LogQueryContext("listUserOrganizationRoles", ctx, d, h)
 
-	// create new Make client
-	c, err := client.GetClient(ctx, d.Connection)
+	c, err := NewMakeClient(d.Connection)
 	if err != nil {
 		return nil, err
 	}
 
 	// prepare params
 	var userId = int(d.EqualsQuals["user_id"].GetInt64Value())
-	var config = client.NewRequestConfig(fmt.Sprintf(`users/%d/user-organization-roles`, userId))
-	if d.QueryContext.Limit != nil {
-		config.Pagination.Limit = int(*d.QueryContext.Limit)
-	}
-
-	// fetch data
-	var pagesLeft = true
-	for pagesLeft {
-		var result = &client.UserOrganizationRoleListResponse{}
-		err = c.Get(&config, result)
+	var uorp = c.NewUserOrganizationRoleListPaginator(int(d.RowsRemaining(ctx)), userId)
+	for uorp.HasMorePages() {
+		organizationRoles, err := uorp.NextPage()
 		if err != nil {
 			plugin.Logger(ctx).Error("make_user_organization_role.listUserOrganizationRoles", "request_error", err)
-			return nil, c.HandleKnownErrors(err, "user:read")
+			return nil, err
 		}
 
-		// stream results
-		for _, i := range result.Users {
+		for _, i := range organizationRoles {
 			d.StreamListItem(ctx, i)
-		}
 
-		// pagination
-		var resultCount = len(result.Users)
-		if d.RowsRemaining(ctx) <= 0 || resultCount < config.Pagination.Limit {
-			pagesLeft = false
-		} else {
-			config.Pagination.Offset += config.Pagination.Limit
+			if d.RowsRemaining(ctx) <= 0 {
+				return nil, nil
+			}
 		}
 	}